+++ modules/Shmuelie.Windows/Shmuelie.Windows.psm1
@@ -30,7 +30,7 @@ if ($IsWindows) {
             [System.Reflection.Assembly]::LoadFrom($dependencyPath) | Out-Null
         }
         Import-Module $appInstallModule -Force -ErrorAction Stop
-        $exportedCmdlets += 'New-AppInstallContext', 'Get-AppInstallSettings'
+        $exportedCmdlets += 'New-AppInstallContext', 'Get-AppInstallItem', 'Get-AppInstallSettings'
     }
 }